// documents/Elevate/src/app/enums/app-content-name.enum.ts
export enum AppContentName {
    Header = 'header',
    HeaderNavigationItem = 'headernavigationitem',
    Image = 'image',
    Post = 'post',
    FinancialSolution = 'financialsolution',
    FinancialSolutions = 'financialsolutions',
    Disclaimer = 'disclaimer',
    KnowledgeCenter = 'knowledgecenter',
    CookieBar = 'customerhubcookiebar'
  }